split up builder.js
