+++ lib/builder.js
@@ -4,10 +4,17 @@ const path = require('path');
 const fs = require('fs');
 const RSVP = require('rsvp');
 const tmp = require('tmp');
-const rimraf = require('rimraf');
 const underscoreString = require('underscore.string');
 const WatchedDir = require('broccoli-source').WatchedDir;
 const broccoliNodeInfo = require('broccoli-node-info');
+
+const NodeWrapper = require('./wrappers/node');
+const TransformNodeWrapper = require('./wrappers/transform-node');
+const SourceNodeWrapper = require('./wrappers/source-node');
+
+const BuilderError = require('./errors/builder');
+const NodeSetupError = require('./errors/node-setup');
+const BuildError = require('./errors/build');
 
 // Clean up left-over temporary directories on uncaught exception.
 tmp.setGracefulCleanup();
@@ -234,9 +241,15 @@ module.exports = class Builder {
     //     in-2 -> ...
     //   02-otherplugin/
     //     ...
-    const tmpobj = tmp.dirSync({ prefix: 'broccoli-', unsafeCleanup: true, dir: this.tmpdir });
+    const tmpobj = tmp.dirSync({
+      prefix: 'broccoli-',
+      unsafeCleanup: true,
+      dir: this.tmpdir,
+    });
+
     this.builderTmpDir = tmpobj.name;
     this.builderTmpDirCleanup = tmpobj.removeCallback;
+
     for (let i = 0; i < this.nodeWrappers.length; i++) {
       const nodeWrapper = this.nodeWrappers[i];
       if (nodeWrapper.nodeInfo.nodeType === 'transform') {
@@ -288,249 +301,18 @@ module.exports = class Builder {
       }
     }
   }
+
+  get features() {
+    return broccoliNodeInfo.features;
+  }
 };
 
 RSVP.EventTarget.mixin(module.exports.prototype);
-module.exports.prototype.features = broccoliNodeInfo.features;
-
-// Base class for builder errors
-class BuilderError extends Error {
-  constructor(a, b, c) {
-    // Subclassing Error in ES5 is non-trivial because reasons, so we need this
-    // extra constructor logic from http://stackoverflow.com/a/17891099/525872.
-    // Note that ES5 subclasses of BuilderError don't in turn need any special
-    // code.
-    let temp = super(a, b, c);
-    // Need to assign temp.name for correct error class in .stack and .message
-    temp.name = this.name = this.constructor.name;
-    this.stack = temp.stack;
-    this.message = temp.message;
-  }
-}
 
 module.exports.BuilderError = BuilderError;
 module.exports.InvalidNodeError = broccoliNodeInfo.InvalidNodeError;
-
-class NodeSetupError extends BuilderError {
-  constructor(originalError, nodeWrapper) {
-    if (nodeWrapper == null) {
-      // Chai calls new NodeSetupError() :(
-      super();
-      return;
-    }
-    originalError = wrapPrimitiveErrors(originalError);
-    const message =
-      originalError.message +
-      '\nat ' +
-      nodeWrapper.label +
-      nodeWrapper.formatInstantiationStackForTerminal();
-    super(message);
-    // The stack will have the original exception name, but that's OK
-    this.stack = originalError.stack;
-  }
-}
 module.exports.NodeSetupError = NodeSetupError;
-
-class BuildError extends BuilderError {
-  constructor(originalError, nodeWrapper) {
-    if (nodeWrapper == null) {
-      // for Chai
-      super();
-      return;
-    }
-
-    originalError = wrapPrimitiveErrors(originalError);
-
-    // Create heavily augmented message for easy printing to the terminal. Web
-    // interfaces should refer to broccoliPayload.originalError.message instead.
-    let filePart = '';
-    if (originalError.file != null) {
-      filePart = originalError.file;
-      if (originalError.line != null) {
-        filePart += ':' + originalError.line;
-        if (originalError.column != null) {
-          // .column is zero-indexed
-          filePart += ':' + (originalError.column + 1);
-        }
-      }
-      filePart += ': ';
-    }
-    let instantiationStack = '';
-    if (originalError.file == null) {
-      // We want to report the instantiation stack only for "unexpected" errors
-      // (bugs, internal errors), but not for compiler errors and such. For now,
-      // the presence of `.file` serves as a heuristic to distinguish between
-      // those cases.
-      instantiationStack = nodeWrapper.formatInstantiationStackForTerminal();
-    }
-    const message =
-      filePart +
-      originalError.message +
-      (originalError.treeDir ? '\n        in ' + originalError.treeDir : '') +
-      '\n        at ' +
-      nodeWrapper.label +
-      instantiationStack;
-
-    super(message);
-    this.stack = originalError.stack;
-
-    // This error API can change between minor Broccoli version bumps
-    this.broccoliPayload = {
-      originalError, // guaranteed to be error object, not primitive
-      originalMessage: originalError.message,
-      // node info
-      nodeId: nodeWrapper.id,
-      nodeLabel: nodeWrapper.label,
-      nodeName: nodeWrapper.nodeInfo.name,
-      nodeAnnotation: nodeWrapper.nodeInfo.annotation,
-      instantiationStack: nodeWrapper.nodeInfo.instantiationStack,
-      // error location (if any)
-      location: {
-        file: originalError.file,
-        treeDir: originalError.treeDir,
-        line: originalError.line,
-        column: originalError.column,
-      },
-    };
-  }
-}
 module.exports.BuildError = BuildError;
-
-class NodeWrapper {
-  constructor() {
-    this.buildState = {};
-  }
-
-  toJSON() {
-    return undefinedToNull({
-      id: this.id,
-      nodeInfo: this.nodeInfoToJSON(),
-      buildState: this.buildState,
-      label: this.label,
-      inputNodeWrappers: this.inputNodeWrappers.map(nw => nw.id),
-      cachePath: this.cachePath,
-      outputPath: this.outputPath,
-      // leave out node, originalNode, inputPaths (redundant), build
-    });
-  }
-
-  formatInstantiationStackForTerminal() {
-    return '\n-~- created here: -~-\n' + this.nodeInfo.instantiationStack + '\n-~- (end) -~-';
-  }
-}
-
 module.exports.NodeWrapper = NodeWrapper;
-
-class TransformNodeWrapper extends NodeWrapper {
-  setup(features) {
-    this.nodeInfo.setup(features, {
-      inputPaths: this.inputPaths,
-      outputPath: this.outputPath,
-      cachePath: this.cachePath,
-    });
-    this.callbackObject = this.nodeInfo.getCallbackObject();
-  }
-
-  build() {
-    const startTime = process.hrtime();
-
-    if (!this.nodeInfo.persistentOutput) {
-      rimraf.sync(this.outputPath);
-      fs.mkdirSync(this.outputPath);
-    }
-
-    return RSVP.resolve(this.callbackObject.build()).then(() => {
-      const now = process.hrtime();
-      // Build time in milliseconds
-      this.buildState.selfTime = 1000 * (now[0] - startTime[0] + (now[1] - startTime[1]) / 1e9);
-      this.buildState.totalTime = this.buildState.selfTime;
-      for (let i = 0; i < this.inputNodeWrappers.length; i++) {
-        this.buildState.totalTime += this.inputNodeWrappers[i].buildState.totalTime;
-      }
-    });
-  }
-
-  toString() {
-    let hint = this.label;
-    if (this.inputNodeWrappers) {
-      // a bit defensive to deal with partially-constructed node wrappers
-      hint += ' inputNodeWrappers:[' + this.inputNodeWrappers.map(nw => nw.id) + ']';
-    }
-    hint += ' at ' + this.outputPath;
-    if (this.buildState.selfTime != null) {
-      hint += ' (' + Math.round(this.buildState.selfTime) + ' ms)';
-    }
-    return '[NodeWrapper:' + this.id + ' ' + hint + ']';
-  }
-
-  nodeInfoToJSON() {
-    return undefinedToNull({
-      nodeType: 'transform',
-      name: this.nodeInfo.name,
-      annotation: this.nodeInfo.annotation,
-      persistentOutput: this.nodeInfo.persistentOutput,
-      needsCache: this.nodeInfo.needsCache,
-      // leave out instantiationStack (too long), inputNodes, and callbacks
-    });
-  }
-}
 module.exports.TransformNodeWrapper = TransformNodeWrapper;
-
-class SourceNodeWrapper extends NodeWrapper {
-  setup(/* features */) {}
-  build() {
-    // We only check here that the sourceDirectory exists and is a directory
-    try {
-      if (!fs.statSync(this.nodeInfo.sourceDirectory).isDirectory()) {
-        throw new Error('Not a directory');
-      }
-    } catch (err) {
-      // stat might throw, or we might throw
-      err.file = this.nodeInfo.sourceDirectory;
-      // fs.stat augments error message with file name, but that's redundant
-      // with our err.file, so we strip it
-      err.message = err.message.replace(/, stat '[^'\n]*'$/m, '');
-      throw err;
-    }
-
-    this.buildState.selfTime = 0;
-    this.buildState.totalTime = 0;
-  }
-
-  toString() {
-    let hint = this.nodeInfo.sourceDirectory + (this.nodeInfo.watched ? '' : ' (unwatched)');
-    return '[NodeWrapper:' + this.id + ' ' + hint + ']';
-  }
-
-  nodeInfoToJSON() {
-    return undefinedToNull({
-      nodeType: 'source',
-      sourceDirectory: this.nodeInfo.sourceDirectory,
-      watched: this.nodeInfo.watched,
-      name: this.nodeInfo.name,
-      annotation: this.nodeInfo.annotation,
-      // leave out instantiationStack
-    });
-  }
-}
 module.exports.SourceNodeWrapper = SourceNodeWrapper;
-
-// Replace all `undefined` values with `null`, so that they show up in JSON output
-function undefinedToNull(obj) {
-  for (const key in obj) {
-    if (obj.hasOwnProperty(key) && obj[key] === undefined) {
-      obj[key] = null;
-    }
-  }
-  return obj;
-}
-
-function wrapPrimitiveErrors(err) {
-  if (err !== null && typeof err === 'object') {
-    return err;
-  } else {
-    // We could augment the message with " [string exception]" to indicate
-    // that the stack trace is not useful, or even set the .stack to null.
-    return new Error(err + '');
-  }
-}
